delete: RealmStorage, MovieRealmDTO 옵셔널 제거

+++ MovieProject/DataBase/MovieRealmData.swift
@@ -30,11 +30,15 @@ extension MovieRealmData {
 
     func saveMovie(movie: Movie?) {
         let movieDTO = movie.map { MovieRealmDTO(movie: $0) }
-        storage.save(movie: movieDTO)
+        if let movieDTO = movieDTO {
+            storage.save(movie: movieDTO)
+        }
     }
 
     func deleteMovie(movie: Movie?) {
         let movieDTO = movie.map { MovieRealmDTO(movie: $0) }
-        storage.delete(movie: movieDTO)
+        if let movieDTO = movieDTO {
+            storage.delete(movie: movieDTO)
+        }
     }
 }
